% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{applyModel}
\alias{applyModel}
\title{applyModel function}
\usage{
applyModel(tY, model, bx)
}
\arguments{
\item{model}{intput}

\item{bx}{intput}

\item{Y}{intput}
}
\value{
data.frame
}
\description{
applyModel function
}
