% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit}
\alias{fit}
\title{fit function}
\usage{
fit(Y, bx, mean.only = FALSE, ref.batch = NULL)
}
\arguments{
\item{Y}{intput}

\item{bx}{intput}

\item{mean.only}{intput}

\item{ref.batch}{intput}
}
\value{
model
}
\description{
fit function
}
