% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{LS.NoRef}
\alias{LS.NoRef}
\title{Main function applying the combat method without a reference batch}
\usage{
LS.NoRef(tY, bx, mean.only = FALSE)
}
\arguments{
\item{bx}{batch indicator variable, coerced to factor}

\item{mean.only}{boolean, if TRUE the Location-only model is applied.}

\item{Y}{input data matrix (observations in rows, spots in columns)}
}
\description{
Main function applying the combat method without a reference batch
}
