% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{LS.Ref}
\alias{LS.Ref}
\title{Main function applying the combat method with a reference batch}
\usage{
LS.Ref(tY, bx, ref.batch = NULL, mean.only = FALSE)
}
\arguments{
\item{bx}{batch indicator variable, coerced to factor}

\item{mean.only}{boolean, if TRUE the Location-only model is applied.}

\item{Y}{input data matrix (observations in rows, spots in columns)}

\item{refbatch}{string identifying the reference batch}
}
\description{
Main function applying the combat method with a reference batch
}
